paypal_mail_to_evm_bot/requirements.txt
python-dotenv
imap-tools
beautifulsoup4
web3 